add b64encode_output option

+++ audio_augmentator/Augmentator.py
@@ -12,6 +12,7 @@ class Augmentator:
 
     def __init__(self,
                  noises_dataset: str,
+                 to_augment: bool = False,
                  decibels: float = 10.0,
                  household_noises: bool = False,
                  pets_noises: bool = False,
@@ -26,6 +27,8 @@ class Augmentator:
                  pre_delay: int = 20,
                  wet_gain: int = 0,
                  wet_only: bool = False):
+
+        self.to_augment = to_augment
         self.to_reverb = to_reverb
         self.noise_decibels = decibels
         self.parameters = {"household_noises": household_noises,
@@ -44,11 +47,13 @@ class Augmentator:
                         wet_gain=wet_gain,
                         wet_only=wet_only))
 
-        self.noises_dataset = DatasetDict.load_from_disk(noises_dataset)
-        self.noise_types = list(self.noises_dataset.keys())
+        if self.to_augment:
+            self.noises_dataset = DatasetDict.load_from_disk(noises_dataset)
+            self.noise_types = list(self.noises_dataset.keys())
 
-    def reverb(self,
-               audio_to_reverb_path: str) -> dict:
+    def reverberate(self,
+                    audio_to_reverb_path: str,
+                    b64encode_output: bool = False) -> dict:
 
         reverbed_result = {}
         filename = Path(audio_to_reverb_path).stem
@@ -64,48 +69,52 @@ class Augmentator:
                                                            channels=1)
                 reverbed_audio_object.export(reverbed_audio_name, format="wav")
                 reverbed_audio_bytes = open(reverbed_audio_name, 'rb').read()
-                reverbed_audio_bytes = b64encode(reverbed_audio_bytes).decode("utf-8")
+                if b64encode_output:
+                    reverbed_audio_bytes = b64encode(reverbed_audio_bytes).decode("utf-8")
                 os.remove(reverbed_audio_name)
                 reverbed_result[reverbed_audio_name] = reverbed_audio_bytes
             except BaseException as err:
                 reverbed_result[reverbed_audio_name] = str(err)
         return reverbed_result
 
-    def augment(self,
-                audio_to_augment_path: str) -> dict:
+    def augmentate(self,
+                   audio_to_augment_path: str,
+                   b64encode_output: bool = False) -> dict:
 
         augmented_audiofiles = {}
         filename = Path(audio_to_augment_path).stem
 
-        for noise_type in self.parameters.keys():
-            augmented_audio_filename = f'{filename}_{noise_type}_{str(int(self.noise_decibels))}.wav'
-            try:
-                if self.parameters[noise_type]:
-                    audio_to_augment, sample_rate = librosa.load(audio_to_augment_path)
-                    audio_to_augment_object = pydub.AudioSegment(data=audio_to_augment.tobytes(),
+        if self.to_augment:
+            for noise_type in self.parameters.keys():
+                augmented_audio_filename = f'{filename}_{noise_type}_{str(int(self.noise_decibels))}.wav'
+                try:
+                    if self.parameters[noise_type]:
+                        audio_to_augment, sample_rate = librosa.load(audio_to_augment_path)
+                        audio_to_augment_object = pydub.AudioSegment(data=audio_to_augment.tobytes(),
+                                                                     sample_width=2,
+                                                                     frame_rate=sample_rate,
+                                                                     channels=1)
+
+                        noises_source = self.noises_dataset[noise_type]
+                        dataset_size = noises_source.num_rows
+                        noise_to_mix_id = random.choice(range(0, dataset_size))
+                        noise_to_mix = noises_source['audio'][noise_to_mix_id]['array'].astype(audio_to_augment.dtype)
+                        noise_to_mix_object = pydub.AudioSegment(data=noise_to_mix.tobytes(),
                                                                  sample_width=2,
                                                                  frame_rate=sample_rate,
                                                                  channels=1)
 
-                    noises_source = self.noises_dataset[noise_type]
-                    dataset_size = noises_source.num_rows
-                    noise_to_mix_id = random.choice(range(0, dataset_size))
-                    noise_to_mix = noises_source['audio'][noise_to_mix_id]['array'].astype(audio_to_augment.dtype)
-                    noise_to_mix_object = pydub.AudioSegment(data=noise_to_mix.tobytes(),
-                                                             sample_width=2,
-                                                             frame_rate=sample_rate,
-                                                             channels=1)
-
-                    amplitude_difference = audio_to_augment_object.dBFS - noise_to_mix_object.dBFS
-                    noise_to_mix_object = noise_to_mix_object.apply_gain(amplitude_difference)
-                    augmented_audio_object = audio_to_augment_object.overlay(
-                        noise_to_mix_object.apply_gain(-self.noise_decibels),
-                        loop=True)
-                    augmented_audio_object.export(augmented_audio_filename, format="wav")
-                    augmented_audio_bytes = open(augmented_audio_filename, 'rb').read()
-                    augmented_audio_bytes = b64encode(augmented_audio_bytes).decode("utf-8")
-                    os.remove(augmented_audio_filename)
-                    augmented_audiofiles[noise_type] = augmented_audio_bytes
-            except BaseException as err:
-                augmented_audiofiles[augmented_audio_filename] = str(err)
+                        amplitude_difference = audio_to_augment_object.dBFS - noise_to_mix_object.dBFS
+                        noise_to_mix_object = noise_to_mix_object.apply_gain(amplitude_difference)
+                        augmented_audio_object = audio_to_augment_object.overlay(
+                            noise_to_mix_object.apply_gain(-self.noise_decibels),
+                            loop=True)
+                        augmented_audio_object.export(augmented_audio_filename, format="wav")
+                        augmented_audio_bytes = open(augmented_audio_filename, 'rb').read()
+                        if b64encode_output:
+                            augmented_audio_bytes = b64encode(augmented_audio_bytes).decode("utf-8")
+                        os.remove(augmented_audio_filename)
+                        augmented_audiofiles[noise_type] = augmented_audio_bytes
+                except BaseException as err:
+                    augmented_audiofiles[augmented_audio_filename] = str(err)
         return augmented_audiofiles
